requirements.txt
opencv-python
numpy
imutils
pyserial